+++ src/main/java/com/tb/Main.java
@@ -1,5 +1,7 @@
 package com.tb;
 
+import com.tb.base.App;
+
 public class Main {
 	public static void main(String[] args) {
 		new App().run();
