+++ src/main/java/com/tb/base/Request.java
@@ -1,4 +1,4 @@
-package com.tb;
+package com.tb.base;
 
 import java.util.HashMap;
 import java.util.Map;
