+++ src/main/java/com/tb/domain/phrase/Phrase.java
@@ -1,4 +1,4 @@
-package com.tb;
+package com.tb.domain.phrase;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
